Refactor single bactraq.py to small code bases
