+++ BacTraq/cli.py
@@ -1,0 +1,81 @@
+import logging
+import argparse
+import pandas as pd
+
+from BacTraq.clustering import cluster_from_distance_matrix
+from BacTraq.cluster_trace import assign_cluster_database
+from BacTraq.reporting import _setup_logging, pretty_name_save, pivot_by_lineage
+
+logger = logging.getLogger('bactraq')
+
+_DEFAULT_THRESHOLD = [20, 10, 5]
+
+def main():
+    parser = argparse.ArgumentParser(
+        description="SNP-distance clustering with consistent naming across runs.",
+        formatter_class=argparse.ArgumentDefaultsHelpFormatter,
+    )
+    parser.add_argument("-d", "--distance-matrix", required=True, metavar="FILE",
+                        help="Path to SNPdist matrix (tab-separated).")
+    parser.add_argument("-o", "--output", required=True, metavar="FILENAME",
+                        help="Output base name. Produces <output>.csv and <output>_history.parquet.gz.")
+    parser.add_argument("-t", "--threshold", metavar="THRESHOLD",
+                        help="Comma-separated SNP thresholds, largest first.\nE.g: --threshold 20,10,5\nDefault: 20,10,5")
+    parser.add_argument("--history", metavar="HISTORY",
+                        help="Path to history .parquet.gz from a previous run.\nGenerate with `bactraq-history`. Omit for clustering only.")
+    parser.add_argument("--nRef", action='store_false',
+                        help="No Reference sample in the SNP distance matrix.")
+    parser.add_argument("--summary", action='store_true',
+                        help="Also write <output>_summary.csv: a cluster-change table (one row per "
+                             "history cluster lineage, Status/New Name per threshold, plus singleton "
+                             "samples) comparing this run against --history. No-op without --history.")
+    args = parser.parse_args()
+
+    distance_matrix = args.distance_matrix
+    save_name = f"{args.output}.csv"
+    log_file = f"{args.output}_bactraq.log"
+    _setup_logging(log_file)
+    logger.info(f"BacTraq run started — input: {distance_matrix}")
+
+    if args.threshold:
+        snp_thres = sorted(map(int, args.threshold.split(',')), reverse=True)
+    else:
+        snp_thres = _DEFAULT_THRESHOLD
+    logger.info(f"Thresholds: {snp_thres}")
+
+    if args.history:
+        logger.info(f"History file: {args.history}")
+        db_cluster = pd.read_parquet(args.history)
+        db_cluster = db_cluster.apply(
+            lambda col: col.map(lambda x: tuple(x) if x is not None else None)
+        )
+    else:
+        logger.info("No history file provided — clustering only, no name tracking")
+        db_cluster = None
+
+    this_run_cluster = cluster_from_distance_matrix(distance_matrix, snp_thres, args.nRef)
+    logger.info(f"Clustering complete — {len(this_run_cluster)} samples")
+
+    change_log = [] if args.summary else None
+    rename_with_db, new_db = assign_cluster_database(this_run_cluster, cluster_db=db_cluster, change_log=change_log)
+
+    pretty_name_save(rename_with_db, distance_matrix, save_name)
+
+    if args.summary:
+        if change_log:
+            changes_df = pd.DataFrame(change_log)
+            level_order = list(this_run_cluster.columns)  # coarsest -> finest
+            summary_table = pivot_by_lineage(changes_df, level_order, current_assignments=rename_with_db, history_assignments=db_cluster)
+            summary_file = f"{args.output}_summary.csv"
+            summary_table.to_csv(summary_file)
+            logger.info(f"Cluster change summary saved to {summary_file}")
+        else:
+            logger.info("--summary requested but no history comparison was available — skipping")
+
+    new_history_file = f'{args.output}_history.parquet.gz'
+    new_db.to_parquet(new_history_file, compression='gzip')
+    logger.info(f"History updated and saved to {new_history_file}")
+    logger.info(f"Log saved to {log_file}")
+
+if __name__=="__main__":
+    main()
